process version number by .create_pkg

+++ .create_pkg.sh
@@ -2,6 +2,13 @@
 
 set -e
 
+VER=$1
+
+if [ "$VER" == "" ]; then
+	echo "no version given"
+	exit 1
+fi
+
 rm -fv ShairTunes2.zip
 
 zip -r ShairTunes2.zip *.pm *.txt *.md *.xml helperBinaries
@@ -9,3 +16,6 @@ zip -r ShairTunes2.zip *.pm *.txt *.md *.xml helperBinaries
 CHK=$(sha1sum -b ShairTunes2.zip|awk '{print $1}')
 
 sed -i "s#<sha>.*</sha>#<sha>${CHK}</sha>#" public.xml
+sed -i "s#version=\"[0-9\.]*\" #version=\"${VER}\" #" public.xml
+
+sed -i "s#<version>0.12</version>#<version>${VER}</version>#" install.xml
